+++ packages/components/src/navigation/item/index.js
@@ -2,11 +2,12 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { noop } from 'lodash';
+import { deburr, noop } from 'lodash';
 
 /**
  * WordPress dependencies
  */
+import { useRef } from '@wordpress/element';
 import { Icon, chevronRight } from '@wordpress/icons';
 
 /**
@@ -14,9 +15,13 @@ import { Icon, chevronRight } from '@wordpress/icons';
  */
 import Button from '../../button';
 import { useNavigationContext } from '../context';
-import { ItemBadgeUI, ItemTitleUI, ItemUI } from '../styles/navigation-styles';
-import { useNavigationTreeItem } from './use-navigation-tree-item';
 import { useNavigationMenuContext } from '../menu/context';
+import { useNavigationTreeItem } from './use-navigation-tree-item';
+import { ItemBadgeUI, ItemTitleUI, ItemUI } from '../styles/navigation-styles';
+
+// @see packages/block-editor/src/components/inserter/search-items.js
+const normalizeInput = ( input ) =>
+	deburr( input ).replace( /^\//, '' ).toLowerCase();
 
 export default function NavigationItem( props ) {
 	const {
@@ -32,13 +37,30 @@ export default function NavigationItem( props ) {
 	} = props;
 	useNavigationTreeItem( props );
 	const { activeItem, setActiveMenu } = useNavigationContext();
-	const { isActive } = useNavigationMenuContext();
+	const { isActive, search } = useNavigationMenuContext();
+	const itemRef = useRef( null );
 
 	// If this item is in an inactive menu, then we skip rendering
 	// We need to make sure this component gets mounted though
 	// To make sure inactive items are included in the navigation tree
 	if ( ! isActive ) {
 		return null;
+	}
+
+	if ( search ) {
+		const searchQuery = normalizeInput( search );
+		if ( title && -1 === normalizeInput( title ).indexOf( searchQuery ) ) {
+			return null;
+		}
+		if (
+			children &&
+			-1 ===
+				normalizeInput( itemRef?.current?.textContent ).indexOf(
+					searchQuery
+				)
+		) {
+			return null;
+		}
 	}
 
 	const classes = classnames( 'components-navigation__item', className, {
@@ -54,7 +76,7 @@ export default function NavigationItem( props ) {
 	};
 
 	return (
-		<ItemUI className={ classes }>
+		<ItemUI className={ classes } ref={ itemRef }>
 			{ children || (
 				<Button href={ href } onClick={ onItemClick } { ...restProps }>
 					{ title && (
